chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "RAG Chatbot for Physical AI & Humanoid Robotics textbook"
authors = ["User <user@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
qdrant-client = "^1.12.0"
cohere = "^5.11.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"